refactored db
Took the common functions to a trait, made admin and user databases sigblings
