+++ Cargo.toml
@@ -17,3 +17,4 @@ rust-argon2 = "0.8"
 rand = "0.7"
 thiserror = "1.0"
 anyhow = "1.0"
+async-trait = "0.1"
