+++ src/db/admin.rs
@@ -1,11 +1,41 @@
-use super::{db, Opt};
-pub use error::Error;
-
-pub type Result<T> = std::result::Result<T, Error>;
+use super::{connect, password, Opt, Result, DB};
 
 /// Administrative database
 pub struct AdminDB {
-    db: db::DB,
+    client: tokio_postgres::Client,
+}
+
+#[async_trait::async_trait]
+impl DB for AdminDB {
+    fn get_name(&self) -> String {
+        "admin".to_string()
+    }
+    fn get_client(&self) -> &tokio_postgres::Client {
+        &self.client
+    }
+    async fn create_all_tables(&self) -> Result<()> {
+        self.get_client()
+            .execute(
+                "CREATE TABLE \"access\" (\"access_type\" TEXT PRIMARY KEY)",
+                &[],
+            )
+            .await?;
+        self.get_client()
+            .execute(
+                "CREATE TABLE \"user\" (\
+                    \"id\" SERIAL PRIMARY KEY,\
+                    \"email\" TEXT NOT NULL UNIQUE,\
+                    \"access\" TEXT NOT NULL,\
+                    \"password_hash\" TEXT NOT NULL,\
+                    FOREIGN KEY(\"access\") REFERENCES \
+                    \"access\"(\"access_type\") \
+                    ON UPDATE CASCADE ON DELETE CASCADE
+                )",
+                &[],
+            )
+            .await?;
+        Ok(())
+    }
 }
 
 impl AdminDB {
@@ -20,14 +50,17 @@ impl AdminDB {
             .password(opt.apiuserpassword.as_str());
         // Connect to the admin database as the default api user
         let admindb = Self {
-            db: db::DB::new(&dbconfig, get_tablespec()).await?,
+            client: connect(&dbconfig).await?,
         };
         // Reset if required
-        if opt.clean && !admindb.db.was_empty() {
-            admindb.db.reset(false).await?;
+        let connected_to_empty = admindb.is_empty().await?;
+        if connected_to_empty {
+            admindb.create_all_tables().await?;
+        } else if opt.clean {
+            admindb.reset_no_backup().await?;
         }
         // Fill access types and the one admin if required
-        if opt.clean || admindb.db.was_empty() {
+        if opt.clean || connected_to_empty {
             admindb.fill_access().await?;
             admindb
                 .insert_admin(
@@ -41,21 +74,12 @@ impl AdminDB {
     /// Fill the access table. Assume that it's empty.
     async fn fill_access(&self) -> Result<()> {
         log::info!("filling presumably empty access table");
-        // Needed entries as json values
-        let access_types: Vec<serde_json::Value> = vec!["admin", "user"]
-            .iter()
-            .map(|t| serde_json::json!(t))
-            .collect();
-        // Entries as a vector of rows
-        let mut access_entries =
-            Vec::<db::RowJson>::with_capacity(access_types.len());
-        for access_type in access_types {
-            let mut access_entry = serde_json::Map::new();
-            access_entry.insert(String::from("access_type"), access_type);
-            access_entries.push(access_entry)
-        }
-        self.db
-            .insert(&db::TableJson::new("access", access_entries))
+        self.client
+            .execute(
+                "INSERT INTO \"access\" (\"access_type\") \
+                VALUES ('admin'), ('user')",
+                &[],
+            )
             .await?;
         Ok(())
     }
@@ -70,24 +94,22 @@ impl AdminDB {
             admin_email,
             admin_password
         );
-        let admin_password_hash = super::password::hash(admin_password)?;
-        let admin_json = format!(
-            "{{\
-            \"email\": \"{}\",
-            \"access\": \"admin\",\
-            \"password_hash\": \"{}\"\
-        }}",
-            admin_email, admin_password_hash
-        );
-        self.db
-            .insert(&db::table::TableJson::new(
-                "user",
-                vec![serde_json::from_str(admin_json.as_str())?],
-            ))
+        let admin_password_hash = password::hash(admin_password)?;
+        self.client
+            .execute(
+                format!(
+                    "INSERT INTO \"user\" \
+                    (\"email\", \"access\", \"password_hash\") \
+                    VALUES ('{}', 'admin', '{}')",
+                    admin_email, admin_password_hash
+                )
+                .as_str(),
+                &[],
+            )
             .await?;
         Ok(())
     }
-    // Authenticates an email/password combination
+    /// Authenticates an email/password combination
     pub async fn authenticate_email_password(
         &self,
         cred: EmailPassword,
@@ -97,9 +119,9 @@ impl AdminDB {
             argon2::verify_encoded(hash.as_str(), cred.password.as_bytes())?;
         Ok(res)
     }
+    /// Returns the password hash for the given email
     async fn get_password_hash(&self, email: &str) -> Result<String> {
         let hash: String = self
-            .db
             .client
             .query_one(
                 "SELECT \"password_hash\" FROM \"user\" WHERE \"email\" = $1",
@@ -111,65 +133,16 @@ impl AdminDB {
     }
 }
 
-/// Tables for the admin database
-fn get_tablespec() -> db::TableSpec {
-    let mut set = db::TableSpec::new();
-    set.push(db::TableMeta::new("access", get_access_colspec(), ""));
-    set.push(db::TableMeta::new(
-        "user",
-        get_user_colspec(),
-        "FOREIGN KEY(access) REFERENCES access(access_type) \
-        ON UPDATE CASCADE ON DELETE CASCADE",
-    ));
-    set
-}
-
-/// Columns for the access table
-fn get_access_colspec() -> db::ColSpec {
-    let mut set = db::ColSpec::new();
-    set.push(db::ColMeta::new("access_type", "TEXT", "PRIMARY KEY"));
-    set
-}
-
-/// Columns for the user table
-fn get_user_colspec() -> db::ColSpec {
-    let mut set = db::ColSpec::new();
-    set.push(db::ColMeta::new("id", "SERIAL", "PRIMARY KEY"));
-    set.push(db::ColMeta::new("email", "TEXT", "NOT NULL UNIQUE"));
-    set.push(db::ColMeta::new("access", "TEXT", "NOT NULL"));
-    set.push(db::ColMeta::new("password_hash", "TEXT", "NOT NULL"));
-    set
-}
-
 #[derive(serde::Deserialize, serde::Serialize)]
 pub struct EmailPassword {
     email: String,
     password: String,
 }
 
-pub mod error {
-    /// Handler errors
-    #[derive(thiserror::Error, Debug)]
-    pub enum Error {
-        /// Argon errors
-        #[error(transparent)]
-        Argon(#[from] argon2::Error),
-        /// Database errors
-        #[error(transparent)]
-        DB(#[from] super::db::Error),
-        /// serde_json errors
-        #[error(transparent)]
-        SerdeJson(#[from] serde_json::Error),
-        /// Represents all cases of `tokio_postgres::Error`
-        #[error(transparent)]
-        TokioPostgres(#[from] tokio_postgres::Error),
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::super::{Opt, StructOpt};
     use super::*;
+    use structopt::StructOpt;
 
     // Create a database
     async fn test_create(clean: bool) -> AdminDB {
@@ -182,8 +155,8 @@ mod tests {
         // Clean or not, there should be one row in the user table
         assert_eq!(
             test_admin_db
-                .db
-                .select("user", &[], "", &[])
+                .get_client()
+                .query("SELECT * FROM \"user\"", &[])
                 .await
                 .unwrap()
                 .len(),
@@ -194,13 +167,15 @@ mod tests {
 
     // Extract first admin's hash
     async fn extract_first_user_hash(db: &AdminDB) -> String {
-        let user_rows = db.db.select("user", &[], "", &[]).await.unwrap();
-        if let serde_json::Value::String(hash) = &user_rows[0]["password_hash"]
-        {
-            String::from(hash)
-        } else {
-            panic!("unexpected lack of string")
-        }
+        db.get_client()
+            .query_one(
+                "SELECT \"password_hash\" FROM \"user\" \
+                WHERE \"id\" = '1'",
+                &[],
+            )
+            .await
+            .unwrap()
+            .get(0)
     }
 
     #[tokio::test]
@@ -209,7 +184,7 @@ mod tests {
         // Start clean
         let test_db = test_create(true).await;
         let hash1 = extract_first_user_hash(&test_db).await;
-        // Restart with backup
+        // Restart
         let test_db = test_create(false).await;
         let hash2 = extract_first_user_hash(&test_db).await;
         assert_eq!(hash1, hash2);
